Add the mingw-w64-busybox package
We already have a BusyBox, of course, in the `busybox` package.

However, that BusyBox is an MSYS2 program, therefore it implicitly links
to msys-2.0.dll, inheriting all the problems of that setup (think: could
not fork). It is necessary, though, because BusyBox uses POSIX
functionality.

Luckily, there is a BusyBox fork called busybox-w32, targeting pure
Win32 API without, say, a fork() syscall.

With a little tweaking, this even builds inside MSYS2.

Let's just package that as a pure MINGW package.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>