git (git-wrapper): when standing in for a builtin, check the case
Misspelling the commant `git status` as `GIT STATUS` will fail due to
the incorrect case: `git.exe` would not think that this is a builtin,
and execute the dashed `git-STATUS.exe`, which in turn would match
`git-status.exe` on Windows (because case-insensitive file system), and
that would be the Git wrapper, which would then call `git.exe STATUS`
again.

This fixes https://github.com/git-for-windows/git/issues/1496.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>