pathtools: add a helper to synchronize all copies
The `mingw-w64-pathtools` directory does not contain a package
definition, on purpose: for ease of maintenance, we do not patch users
of this code to link to a library, but we patch in the source code
directly.

Which means that we have to maintain copies of this code, and after
every update to `mingw-w64-pathtools/pathtools.[ch]`, we have to
synchronize those copies and increment pkgrels.

Let's add a simple shell script to do that job for us.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>