Add a GitHub workflow to build and deploy a single MINGW package
This is the equivalent to
https://github.com/git-for-windows/MSYS2-packages/blob/main/.github/workflows/build-and-deploy.yml
which Git for Windows had to introduce in order to provide backports for
i686 variants of the MSYS packages (MSYS2 no longer maintains those).

We now need to build and deploy a MINGW package manually because we need
it a fixed mingw-w64-gettext really quickly so that we can build and
release Git for Windows v2.35.0-rc0.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>