TEST: git: force-drop obsolete compatibility option
In https://github.com/git-for-windows/git/commit/fa93bb20d72e (MinGW:
Fix stat definitions to work with MinGW runtime version 4.0,
2013-09-11), the `-D_USE_32BIT_TIME_T` flag was force-set in Git for
Windows.

Since we're no longer targeting MinGW, but mingw-w64 instead, this is no
longer necessary. It is even undesirable because it causes errors in
UCRT64, which is incompatible with this flag.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>