git: verify the completeness of the `max_allowed_tree_depth` hack
In git/git@436a42215e5 (max_tree_depth: lower it for clangarm64 on
Windows, 2025-04-23), I worked around a vexing, ARM64-specific bug in
the tree-depth restriction where Git ran into a ``STATUS_STACK_OVERFLOW`
that Git simply could not handle correctly.

This bug might not actually be limited to the specific circumstances for
which I instated that work-around. Let's verify as part of the build
that this work-around is complete enough as far as MSYS2 is concerned.

Technically, this should live in the `check` function, not in the
`build` function. But in MSYS2's PullRequests, the `check` part is
deliberately omitted, so we cannot use that approach.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>