TEST: git: switch to my development branch of git-for-windows/git
There are now a good deal of work-arounds for things in
git-for-windows/git that need to be adjusted before we can mark
https://github.com/msys2/MINGW-packages/pull/26470 as ready for review.
Better develop the patches directly in the Git repository than playing
more of those `config.mak` and `sed` games.

This here commit will of course be dropped as soon as things work and
the git-for-windows/git patches made it into a tagged version of Git for
Windows.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>